+ General code quality improvements. Thanks to James for letting me know of some of these coding conventions.
+ StarBlockCache::deleteBlocks() and StarBlockCache::useBlock() must
also free the memory alloted for StarBlock::star

CCMAIL: kstars-devel@kde.org


svn path=/branches/kstars/summer/kdeedu/kstars/; revision=818557

+++ kstars/skycomponents/starblockcache.h
@@ -103,7 +103,7 @@ class StarBlockCache {
      */
     int deleteBlocks(int nblocks);
 
-    long useID;                // A number identifying the current 'use' cycle
+    quint32 useID;             // A number identifying the current 'use' cycle
 
  private:
